+++ R/largeVis.R
@@ -20,10 +20,11 @@
 #' @return A `largeVis` object with the following slots:
 #'  \describe{
 #'    \item{'knns'}{An [N,K] 0-indexed integer matrix, which is an adjacency list of each vertex' identified nearest neighbors.
-#'    If the algorithm failed to find \code{K} neighbors, the matrix is padded with \code{NA}'s.}
+#'    If the algorithm failed to find \code{K} neighbors, the matrix is padded with \code{NA}'s. Note that this matrix is not identical to the output
+#'    from \code{\link{randomProjectionTreeSearch}}: missing neighbors are \code{NA}'s rather than \code{-1}'s, and the matrix is transposed.}
 #'    \item{'wij'}{A sparse [N,N] matrix where each cell represents \eqn{w_{ij}}.}
 #'    \item{'call'}{The call.}
-#'    \item{'coords'}{A [N,D] matrix of the embedding of the dataset in the low-dimensional space.}
+#'    \item{'coords'}{A [D,N] matrix of the embedding of the dataset in the low-dimensional space.}
 #'  }
 #'
 #' @export